requirements.txt
yt-dlp>=2022.7.18
pyqt6==6.3.1
pytest>=7.1.2,<7.1.3
python-magic>=0.4.27,<0.4.28
PyYAML>=6.0,<6.1